scripts/requirements.txt
click>=8.0
rich>=13.0
pyyaml>=6.0